+++ dtoolcore/__init__.py
@@ -3,6 +3,7 @@
 """
 
 import os
+import sys
 
 import datetime
 import logging
@@ -11,7 +12,6 @@ import shutil
 import tempfile
 import uuid
 
-from pkg_resources import iter_entry_points
 from collections import defaultdict
 
 import dtoolcore.utils
@@ -59,7 +59,19 @@ def _generate_storage_broker_lookup():
     """Return dictionary of available storage brokers."""
     logger.debug("In _generate_storage_broker_lookup...")
     storage_broker_lookup = dict()
-    for entrypoint in iter_entry_points("dtool.storage_brokers"):
+
+    if sys.version_info >= (3, 8):
+        from importlib.metadata import entry_points
+        eps = entry_points()
+        if sys.version_info >= (3, 10):
+            entrypoints = eps.select(group="dtool.storage_brokers")
+        else:
+            entrypoints = eps.get("dtool.storage_brokers", [])
+    else:
+        from pkg_resources import iter_entry_points
+        entrypoints = iter_entry_points("dtool.storage_brokers")
+
+    for entrypoint in entrypoints:
         StorageBroker = entrypoint.load()
         key = StorageBroker.key
         logger.debug("_gnerate_stroage_broker_lookup.key: {}".format(key))